Clear all login sessions on permissions change
